Honor content-length set on request body
This commit ensures that when a RequestBody is created with an InputStream, the
content-length also set on the RequestBody is honored, regardless of how much
extra data is available in the stream. Note that if less data is actually
available in the wrapped stream, this implmenetation will return EOF as well.

Fixes #2908