Implement retryable() method in exceptions.
WHAT?

Implement the `retryable()` method in the `AwsServiceException` and
`SdkServiceException` classes.

WHY?

The SDK internally has a way to determine whether an exception is
retryable based on the status code and the error code. This change uses
this existing logic to implement the `retryable()` method of the
exception classes so that it no longer returns a static `false` value.

HOW?

Successful build of `./mvnw clean install -pl :aws-core -am`