Fix issue in IMDS credentials provider that causes expired credentials to be vended for a short period of time after the credentials provider is inactive for a long time. (#3314)
Before this change, if credentials are stale (because they weren't prefetched, likely due to inactivity) only one thread would block to refresh and other calling threads would be given expired credentials. This is good during an IMDS outage if the credential expiration has been extended service-side, but it's bad when the credentials are actually expired.

After this change, if credentials are stale and we go to refresh them, we'll hold other calling threads until that refresh completes. This will cause increased latency during credential refreshes during an IMDS outage, but ensure that vending expired credentials is minimized outside of outage scenarios.