Generate rules create at build + fixes (#3399)
The primary change in this commit is to generate the creation of the
EndpointRuleSet as a static method in the default provider rather than loading
the rule set from JSON at runtime. This means we don't incur any disk I/O cost
at runtime and possibly even remove the dependency on json-util later.

There are also some other minor fixes included in this PR including a working
resolveEndpoint() implementation on the default provider.