Validate signer client options based on authtype
This fixes an issue with the client option validation where the client builder
is incorrectly validating that a value for the SIGNER option is present, even
when it's not required (because the service does not use AWS auth).

This change only does the appropriate validations based on the authtypes present
in the service model.