Make Jackson builds reproducible
Unless we actually update the Jackson versions, the only differences in the JARs
between versions are trivial: differences in the pom.xml, pom.properties, and
metadataa entries in MANIFEST.MF, all of which are not necessary.

This removes those pieces of information from the JARs so that they are
identical.