Codegen S3CrtAsyncClientBuilder
This ensures that new additions to S3's set of ClientContextParams are always
accounted for. Note that the implementation is still handwritten so when S3 adds
new ones, we would still need to make a small code change, but at least there
will be a clear signal since the build will fail.