Add new ErrorType metric
This commit adds a new ErrorType core metric that is recorded for all failed API
call attempts. The ErrorType records the general category  of error that ocurred
for a failed API call attempt. Those categories are:

 - Throttling errors
 - Service errors other than throttling
 - I/O errors
 - API call or API call attempt timeouts

The intent of this metric is to help locate possible issues at a glance and help
direct further debugging or investigation.