Fixed bug where using profile-based credentials could cause the SDK to read the profile file with each request.
This bug was introduced with the profile file supplier changes. If the customer was not overriding the profile file via the client override config AND was relying on profile file-based configuration, that configuration would be read with every request. This had a significant performance cost.

```
Benchmarking Results (software.amazon.awssdk.benchmark.apicall.protocol.JsonProtocolBenchmark)
Before:   2858.805 ± 181.491  ops/s
After:    10915.079 ± 677.022  ops/s
```