Fix leaked futures when executor rejects Runnable (#4009)
This commit fixes  an issue where CompletableFutures are leaked/never completed
when the submission to the FUTURE_COMPLETE_EXECUTOR is rejected.

By default, the SDK uses 2 * number of cores (with a maximum of 64), and uses
bounded queue of size 1000. In cases where the throughput to the client exceeds
the executor's ability to keep up, it would reject executions, and previously,
this would lead to leaked futures.