2/2 Improve HTTP2/PING timeout logic
Improve PING timeout logic, by allowing data frame packets to reset the timer.
Since the goal is to identify and tear down a dead connection, receiving
data packets serve as activity as well. This will help in case of scheduling
and network delays if the PING ACK is not received in time.