Remove dependency on http-auth-aws from codecatalyst.
Previously, AwsEndpointProviderUtils and AuthSchemeUtils, which are copied into every client, depend on classes from http-auth-aws. As a result, every client needed a dependency on http-auth-aws. With this change, we move the methods from those classes that depend on http-auth-aws to be code generated within only the clients that need them.