Make EventBridgeMultiRegionEndpointIntegrationTest reliable on the first run.
Previously, the test would wait for the endpoint state to be active, but this did not mean that the SDK could resolve the DNS entry for the endpoint. Now, we wait for the DNS entry to be resolvable. This won't appreciably affect the time it takes to run the test, but can avoid false positive failures when the test is run on a new account for the first time.