Do not short-circuit toSdkHttpFullRequest
With the short-circuit it is possible that the signedRequest.payload()
isn't set on the resulting SdkHttpFullRequest. In cases where the
payload doesn't actually change during HttpSigner.sign(), this is
ok. But the payload *can* change, e.g., for chunk-encoding, so need to
extract the resulting payload from signedRequest.payload().

This change was already made to SigningStage.