Enable auto reads on idling connections (#4655)
This change enables auto reads on idling connections, and disables them on
in-use connections (no auto read for in-use connections is existing behavior).
Auto reads while the channel is idling allows the channel to read and
immediately handle events that are triggered by the remote end such as TLS close
notifies. In particular this fixes issues where a channel is leased only for the
SDK to find that the remote end has initiated a shutdown of the TLS connection.