Support using SLF4J implementations for logging in the bundle
The bundle shades the SLF4J classes. By adding a bridging mechanism, this allows
logging through SLF4J to continue to work even when using the bundle.

To get around the shading, the SDK bundle is moved to `bundle-sdk`, so that we
can include the `bundle-logging-bridge` artifacts in the final `bundle` artifact
without the the SLF4J classes being shaded.