Reduce memory usage in S3 when plugins aren't used.
Modifying the SDK client configuration at the request level uses extra memory, because it requires copying the existing configuration map.

Before this change, we always modify the client configuration with each request to add the client instance. This change moves this modification to the client-creation, so that the configuration doesn't need to be copied unless plugins also modify the configuration.

This change also removes the conditional logic that only added the client for S3, which simplifies the code generator.