Fix check for release tag
The original check checks whether the tag *exists* in the local and then tries
to remove it from the public repo. However, it's possible that the tag is in the
private repo, where it might not exists (refs from both remotes are present
locally).

This commit fixes this by explicitly check if the tag exists in the public repo
and only then will it try to delete it.