Choose next potential scheme if signer not found
This commit fixes the generated auth scheme selection mechanism by
ensuring that auth scheme is able to provide a signer implementation
(i.e. does not throw an exception) before selecting the auth scheme.
This is useful in situations where the signer implementation relies on
components thay may not be available at runtime, such as the
AwsV4aAuthScheme, which relies on CRT which is an optional dependency.

Additional changes:
 - introduce crt-unavailable-tests which is a good candidate for testing
   behavior of components that use CRT when CRT is not present.
 - modify DefaultAwsV4aAuthScheme so that errors thrown during static
   initialization of the singleton holder for the Sigv4a signer are
   caught and thrown directly instead of resulting in
   ExceptionInInitializerError