Consume any pending bytes on socket
This update checks the wrapped socket to see if there is any data
available on the socket. If this is true, then there is (at least a
partial) TLS record that is waiting to be read. Other than application
data, this record may contain a close_notify.

If data is available, then we perform a single-byte read on the SSL
socket to get the SSLSocket to consume the record.