Fix test failures
Http2PingHandlerTest - Since the PING is now only seen
from PingHandler to the tail of the pipeline, need to
move pingResponder and Write simulator handler so it
can see the ping packets

PingTimeoutTest - Since Idle reaper never worked before
the test did not consider that, need to make sure
the idle timer is long enough so the ping test
connection does not get closed so ping behavior
is validated.

SubscribeToShardIntegrationTest - activity on stream level
is not observed by the idle handler, and it needs to be moved
before the multiplex handler which moves the traffic to child
stream channel pipeline. However this is not sufficient
and it requires Connection pooling to mark the underlying
connection as in-use instead of the stream for the reaper
to not reap connections that are being used, this is necessary
especially for any request which could take 5 seconds or more
to get a response from server, so the connection does not
get reaped while it is in use.