Stop creating anonymous HashMap in models (#5748)
This commit updates the generation of the service models classes so that
we don't create anonymous HashMaps for the `SDK_NAME_TO_FIELD` field.
Rather, we move the initialization to a static method that populates a
normal HashMap returns it. This lowers the size of the JARs because the
compiler doesn't need to create separate .class files for each of these
anonymous implementations of HashMap.