Fixed an issue that could cause checksum mismatch errors in S3 uploads. (#5836)
* Fixed an issue that could cause checksum mismatch errors in S3 uploads.

This error is sometimes encountered when a customer uses:
1. The AsyncS3Client
2. A ChecksumAlgorithm of SHA1 or SHA256 (instead of the default CRC32)
3. Parallel uploads

The root cause was the SDK using thread locals to cache the SHA1 or SHA256 message digest implementations. This meant that if a single event loop thread was processing multiple requests, those two requests would use the same digest implementation to calculate the checksum.

* Fixed an issue where SignerUtils was using the internal class of another module.

* Update clone() failure message. Move to LinkedBlockingDeque for DigestAlgorithm cache.

* Added direct testing of the DigestAlgorithm cache.

* Fixed a transient issue that could occur when the first attempt fully sends the payload, and then a retry happens with checksums enabled.