Validate sent body len, remove buffering
This commit removes the buffering of
`RequestBody.fromContentProvider(ContentStreamProvider,long,String)`,
which was used to ensure that the provider always provided the expected
number of bytes given by `length`.

Instead of doing the buffering, this commit updates
`SdkLengthAwareInputstream` to validate the number of bytes read is
correct, and throws an error if EOF is reached prematurely. We use this
stream to ensure that we sent the correct amount of bytes to the server
when sending the HTTP request.

Note that this a weaker check because we're only testing for the
*number* of bytes sent rather than their contents but has the benefit of
sidestepping buffering.