Improve the IP addresses detection (#5911)
* Improve the IP addresses detection

The previous code relied on the exception thrown by `Integer.parseInt`
to detect when the given segment was not a integer number. This uses
exception as control flow and it can be very inefficient as creating
an exception is costly.

Instead of using `Integer.parseInt` we just handle the few cases that
we care about depending on the length of the segment. If the length is
one then check that the char is a digit, if the length is two then
check that both chars are digits, if the length is 3, check that all
the chars are digits and that the converted number is less than 255.

* Address PR comments

* Remove unused imports