Remove def 16MB read chunk size
This commit removes the hardcoded 16MiB read chunk size for uploadFile()
in the TransferManager. This was originally a workaround to improve
performance when using CRT; however this is no longer necessary because
file uploads are done entirely in CRT without the SDK/Java layer doing
any reads.

Without this default size, the default read chunk size reverts back to
the default from `FileAsyncRequestBody` which is 16KiB.