Update rewrite versions at build time (#5931)
* Update rewrite versions at build time

This change changes add-transfer-manager-dependency.yml and
upgrade-sdk-dependencies.yml to velocity templates so that the version
number can be set at build time when buliding the v2-migration module.
This moves this version update process out of the buildspecs so that we
don't have to rely on them to have been run for builds to work
correctly.

* Move version to a property