Add retry tests for request bodies (#6056)
* Add retry tests for request bodies

This commit adds tests for implementations of `RequestBody` and
`AsyncRequestBody` where the requests are retried to ensure that they send the
same content for every attempt.

* Add 0 byte test

* Try to reduce memory usage

* Switch to plain Jetty server

Unfortunately WireMock saves all the request bodies which increases
memory usage. Use a plain Jetty server that doesn't save requests to
avoid this. We don't rely on any WireWock request assertions so it works
out.

* Lower max size

* Max 8MB

* Force s3-tests to run at end

* Increase memory