Add support for validating models in generator
This commit introduces `ModelValidator` which allows us to do validation
of the `IntermediateModel` before generation starts; violations are
emitted to a `validation-report.json` file under
`generated-sources/sdk/models`.

As part of this change, `CodeGenerator` was updated to accept
`IntermediateModel` as a parameter. This is because all validators
operate on the `IntermediateModel`, which already have any SDK specific
customizations applied. Since some validators require referencing
multiple models, (such as when we start validating shared shapes),
having the ability to pass in `IntermediateModel` directly means we only
need to build these models once, upfront.