Add validation for .brazil.json
This commit adds the `validate-brazil-config` script, which validates
the .brazil.json file for correctness. It validates that all modules of
the SDK are mapped to a package internally, or explicitly skipped.
Additional, it ensures that all external dependencies are also mapped to
internal packages.

The buildspecs/validate-brazil-config.yml is so that this validation can
run as PR build check.