Fix classloading in ConstructorCache (#6233)
Use the SDK standard ClassLoaderHelper#getClass to try and load the
given class from a ClassLoader.

Previously, we were only looking at the ClassLoader returned by
ClassLoaderHelper#contextClassLoader which either returns the Thread's
context ClassLoader, or the system ClassLoader if that isn't present.
This is an issue because it's possible that neither of these
ClassLoaders are correct; we should also be looking at the ClassLoader
that loaded the calling class.

As part of this change, the internal Map used in ConstructorCache has
been changed. Rather than doing a two-level mapping from String -> (Map
of ClassLoader to Class), we have a single level mapping from String ->
Class. Class is still referenced via a WeakReference so it is still free
to be GC'd at any point.