Use actual member name for NULL
In DynamoDB, the "NULL" member of AttributeValue is customized to "NUL"
because "null" is a reserved keyword. This has the effect of generating
nul() and getNul() as getters, and nul(Boolean), and setNul(Boolean) as
setters for the serializable builder class.

If using the serializable builder class to map from a JSON string with
the correct property name "NULL", it will not be recognized.

To fix this, generate an additional setter with the correct name
getNull().

To make this work, `ModifyModelShapeModifier` has been updated with a
new option called `alternateBeanPropertyName`. When set, for the
modified member, we will generate an additional bean style setter on the
model builder using the given property name.