Don't rely on endpoint params region
When generating the for the endpoint interceptor, don't assume that the
endpoint params contains a 'region' member; endpoint rulesets are not
required to have this parameter; in situations where the endpoint params
does not have this member, compilation fails.

For backwards compatibility, give the region endpoint params preference
if it exists; otherwise use the region from the execution attributes.