Merge feature/master/checksum-reuse (#6452)
* Reuse computed checksums across retries (#6413)

* Reuse computed checksums across retries

This commit adds the ability to reuse previously computed checksums for
a request across retries.

This ensures that if a request data stream is modified between attempts
that the server will reject the request.

As part of this change, the `http-auth-spi` package has been updated to
expose a new interface: `PayloadChecksumStore`. This is a simple storage
interface that allows signers to store and retrieve computed checksums.
Additionally, a new `SignerProperty` is introduced,
`SdkInternalHttpSignerProperty.CHECKSUM_CACHE` so that signers and
their callers can access this cache.

Note that both the interface and associated signer property are
`@SdkProtectedApi` and not intended to be used by non-SDK consumers of
`http-auth-spi`.

Finally, this adds a dependency on `checksums-spi` for `http-auth-spi`.

* Update core/sdk-core/src/main/java/software/amazon/awssdk/core/internal/http/pipeline/stages/SigningStage.java

Co-authored-by: David Ho <70000000+davidh44@users.noreply.github.com>

* Review comments

---------

Co-authored-by: David Ho <70000000+davidh44@users.noreply.github.com>

* Reuse computed checkums for async payloads (#6441)

* Reuse computed checkums for async payloads

This is a continuation of #6413 that adds support for checksum reuse on
async payloads.

Additionally, this commit renames references to "ChecksumCache" to
"ChecksumStore".

* Fix cross module issue

* Reuse checksums in legacy signing codepath (#6444)

* Reuse checksums in legacy signing codepath

This commit adds support for reusing calculated payload checksums over
retries in the legacy (i.e. non-SRA) signing codepaths.

* S3 testing with non-SRA

* Add changelog (#6454)

* Don't store x-amz-content-sha256 (#6458)

This doesn't correspond to a flexible checksum value, so this is not
required.

---------

Co-authored-by: David Ho <70000000+davidh44@users.noreply.github.com>