Auto add changelog moves
The changelogs added to .changes/<previous minor version> and
changelogs/<previous minor version>.md are new files that are not
tracked by Git and it's easy to miss doing a 'git add' for them when
running scripts/minor-version-bump. Add them automatically so it's
ensured that they're part of the version bump commit.