Don't special case 0 TTL in Apache 5 (#6606)
* Define infinite TTL as <= 0

This commit updates the definition of the CONNECTION_TIME_TO_LIVE (TTL)
so that non-positive values indicate an infinite TTL. The reason for
this is that in contrast to Apache 4.x, Apache 5.x uses 0 to indicate
immediate expiry. This updated definition allows us to use a new default
value of -1 across all supported HTTP clients to indicate an infinite
TTL.

Note that in practice, this is already the case, as all clients other
than the Apache 5.x client treat non-positive values as infinite:

 - For Apache 4.x, value of <= 0 is treated as infinite: https://github.com/apache/httpcomponents-core/blob/a5c117028b7c620974304636d52f06f172f1d08b/httpcore/src/main/java/org/apache/http/pool/PoolEntry.java#L87-L93
 - For Netty, `OldConnectionReaperHandler` is only used when the TTL
   value > 0, otherwise the handler is not added to the channel's
   pipeline:
   https://github.com/aws/aws-sdk-java-v2/blob/90f17ab300e27bee5367e4eff30b49bfa1f06af0/http-clients/netty-nio-client/src/main/java/software/amazon/awssdk/http/nio/netty/internal/ChannelPipelineInitializer.java#L213-L216
 - CRT and URLConnection client does not support this configuration.

* Don't special case 0 TTL in Apache 5

Shadow the global default CONNECTION_TIME_TO_LIVE value of 0 to -1
within the Apache 5 client instead of special casing the 0 value when
building the connection config.

* Revert "Define infinite TTL as <= 0"

This reverts commit 86a120a818ebe40426412c58160587aaf9f27126.

* Add tests

* Add changelog

* Remove check

* Review comments

* Update javadoc,changelog